+++ package/setup/views/example/index.blade.php
@@ -1,4 +1,7 @@
-<p>You can find the files for this setup at:</p>
+<h3>Index</h3>
+<p>This is the index method for your installed <code>ExampleController</code>.</p>
+<p>All the items in the navigation on the left are controllers and methods from different folders; you click them, they run :)</p>
+<p>The files for this setup can be found at:</p>
 <ul>
 	<li><code>{{ $controllers }}</code></li>
 	<li><code>{{ $views }}</code></li>