+++ package/setup/views/example/welcome.blade.php
@@ -1,2 +1,2 @@
-<h1 style="color: #18bc9c;">Hello, {{ $name }}!</h1>
+<h3 style="color: #18bc9c;">Hello, {{ $name }}!</h3>
 <p>Edit the field above to change the output...</p>