+++ package/views/help/basics/run.blade.php
@@ -1,0 +1,3 @@
+<p>This is a method; it was called at:</p>
+<pre>{{ $date }}</pre>
+<p>Run it again by clicking the <strong>Run</strong> button</p>